Proseguimento testing d'unità

+++ eSport/src/model/ProdottoModel.java
@@ -24,9 +24,21 @@ public class ProdottoModel {
 	 * @throws SQLException 
 	 */
 	public void doSave(ProdottoBean prodotto) throws SQLException {
+		log.info("ProdottoModel -> doSave");
+
 		Connection connection=null;
 		PreparedStatement preparedStatement=null;
 		
+		log.info("doSave -> verifico pre-condizioni");
+		if(prodotto==null || prodotto.getCodice()==null || prodotto.getCodice().equals("")
+				|| prodotto.getNome()==null || prodotto.getNome().equals("")
+				|| prodotto.getMarca()==null || prodotto.getMarca().equals("")
+				|| prodotto.getTipo()==null || prodotto.getTipo().equals("")
+				|| prodotto.getDescrizione()==null || prodotto.getDescrizione().equals("")
+				|| prodotto.getQt()<1 || prodotto.getIva()<1 || prodotto.getPrezzo()<1)
+			return;
+
+		log.info("doSave -> eseguo query");
 		String insertSQL="insert into " + ProdottoModel.TABLE_NAME
 				+ " (codice, nome, tipo, marca, qt, prezzo, iva, descrizione) "
 				+ "values (?, ?, ?, ?, ?, ?, ?, ?)";
@@ -57,6 +69,7 @@ public class ProdottoModel {
 				DriverManagerConnectionPool.releaseConnection(connection);
 			}
 		}
+		log.info("ProdottoModel -> doSave terminato");
 	}
 	
 	/**
@@ -66,12 +79,19 @@ public class ProdottoModel {
 	 * @throws SQLException 
 	 */
 	public Set<ProdottoBean> doRetrieveByTipo(String tipo, String order) throws SQLException{
+		log.info("ProdottoModel -> doRetrieveByTipo");
+		
 		LinkedHashSet<ProdottoBean> prodotti=new LinkedHashSet<ProdottoBean>();
 		TagliaModel tagliaModel=new TagliaModel();
+		
+		log.info("doRetrieveByTipo -> verifico pre-condizioni");
+		if(tipo==null || tipo.equals(""))
+			return null;
 		
 		Connection connection=null;
 		PreparedStatement preparedStatement=null;
 		
+		log.info("doRetrieveByTipo -> eseguo query");
 		String selectSQL="select * from " + ProdottoModel.TABLE_NAME + " where tipo=?";
 		
 		if (order!=null && !order.equals("")) {
@@ -110,6 +130,7 @@ public class ProdottoModel {
 				DriverManagerConnectionPool.releaseConnection(connection);
 			}
 		}
+		log.info("ProdottoModel -> doRetrieveByTipo terminato");
 		
 		return prodotti;
 	}
@@ -121,12 +142,19 @@ public class ProdottoModel {
 	 * @throws SQLException 
 	 */
 	public ProdottoBean doRetrieveByCodice(String codice) throws SQLException{
+		log.info("ProdottoModel -> doRetrieveByCodice");
+
 		ProdottoBean bean=new ProdottoBean();
 		TagliaModel tagliaModel=new TagliaModel();
+		
+		log.info("doRetrieveByCodice -> verifico pre-condizioni");
+		if(codice==null || codice.equals(""))
+			return null;
 		
 		Connection connection=null;
 		PreparedStatement preparedStatement=null;
 		
+		log.info("doRetrieveByCodice -> eseguo query");
 		String selectSQL="select * from " + ProdottoModel.TABLE_NAME + " where codice=?";
 
 		try {
@@ -157,6 +185,7 @@ public class ProdottoModel {
 				DriverManagerConnectionPool.releaseConnection(connection);
 			}
 		}
+		log.info("ProdottoModel -> doRetrieveByCodice terminato");
 		
 		return bean;
 	}
@@ -167,11 +196,24 @@ public class ProdottoModel {
 	 * @throws SQLException 
 	 */
 	public boolean doUpdate(ProdottoBean prodotto) throws SQLException {
+		log.info("ProdottoModel -> doUpdate");
+
 		Connection connection = null;
 		PreparedStatement preparedStatement = null;
 
+		log.info("doUpdate -> verifico pre-condizioni");
+		if(prodotto==null || prodotto.getCodice()==null || prodotto.getCodice().equals("")
+				|| prodotto.getNome()==null || prodotto.getNome().equals("")
+				|| prodotto.getMarca()==null || prodotto.getMarca().equals("")
+				|| prodotto.getTipo()==null || prodotto.getTipo().equals("")
+				|| prodotto.getDescrizione()==null || prodotto.getDescrizione().equals("")
+				|| prodotto.getQt()<1 || prodotto.getIva()<1 || prodotto.getPrezzo()<1
+				|| doRetrieveByCodice(prodotto.getCodice())==null)
+			return false;
+		
 		int result=0;
 
+		log.info("doUpdate -> eseguo query");
 		String updateSQL="update " + ProdottoModel.TABLE_NAME + " "
 					   + " set nome=?, "
 					   + " marca=?, "
@@ -194,6 +236,8 @@ public class ProdottoModel {
 			preparedStatement.setString(7, prodotto.getCodice());
 			
 			result=preparedStatement.executeUpdate();
+			
+			connection.commit();
 		} 
 		finally {
 			try {
@@ -204,6 +248,7 @@ public class ProdottoModel {
 				DriverManagerConnectionPool.releaseConnection(connection);
 			}
 		}
+		log.info("ProdottoModel -> doUpdate terminato");
 		
 		return (result!=0);
 	}
@@ -214,11 +259,23 @@ public class ProdottoModel {
 	 * @throws SQLException 
 	 */
 	public boolean doDelete(ProdottoBean prodotto) throws SQLException {
+		log.info("ProdottoModel -> doDelete");
 		Connection connection = null;
 		PreparedStatement preparedStatement = null;
 
+		log.info("doDelete -> verifico pre-condizioni");
+		if(prodotto==null || prodotto.getCodice()==null || prodotto.getCodice().equals("")
+				|| prodotto.getNome()==null || prodotto.getNome().equals("")
+				|| prodotto.getMarca()==null || prodotto.getMarca().equals("")
+				|| prodotto.getTipo()==null || prodotto.getTipo().equals("")
+				|| prodotto.getDescrizione()==null || prodotto.getDescrizione().equals("")
+				|| prodotto.getQt()<1 || prodotto.getIva()<1 || prodotto.getPrezzo()<1
+				|| doRetrieveByCodice(prodotto.getCodice())==null)
+			return false;
+		
 		int result=0;
 
+		log.info("doDelete -> eseguo query");
 		String deleteSQL="delete from " + ProdottoModel.TABLE_NAME + " where codice=?";
 
 		try {
@@ -227,6 +284,8 @@ public class ProdottoModel {
 			preparedStatement.setString(1, prodotto.getCodice());
 
 			result=preparedStatement.executeUpdate();
+			
+			connection.commit();
 		} 
 		finally {
 			try {
@@ -237,6 +296,7 @@ public class ProdottoModel {
 				DriverManagerConnectionPool.releaseConnection(connection);
 			}
 		}
+		log.info("ProdottoModel -> doDelete terminato");
 		
 		return (result!=0);
 	}
